[157000787] Refactor Exporter Job Properties
- Refactor exporter job properties to separate all from 'public only' submissions output
- Move configuration related classes to a separated package
