+++ Exporter/src/main/java/uk/ac/ebi/biostd/exporter/jobs/full/configuration/FullExportAllSubmissionsProperties.java
@@ -1,0 +1,7 @@
+package uk.ac.ebi.biostd.exporter.jobs.full.configuration;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "jobs.full.allSubmissions")
+public class FullExportAllSubmissionsProperties extends FullExportFileProperties {
+}
