+++ Exporter/src/main/java/uk/ac/ebi/biostd/exporter/jobs/full/configuration/FullExportFileProperties.java
@@ -1,0 +1,11 @@
+package uk.ac.ebi.biostd.exporter.jobs.full.configuration;
+
+import lombok.Data;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+public class FullExportFileProperties {
+    private String fileName;
+    private String filePath;
+}
